requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.0
altair==4.2.2
altgraph==0.17.2
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
blosc2==2.0.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.0
deepl==1.14.0
docker==6.1.2
enchant==0.0.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fasttext==0.9.2
Flask==2.3.2
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
keras-tuner==1.3.5
kiwisolver==1.4.4
kt-legacy==1.0.5
langcodes==3.3.0
libclang==16.0.0
llvmlite==0.40.0
macholib==1.15.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mlflow==2.3.2
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.7
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.1.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyaml==23.5.9
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
querystring-parser==1.2.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
six==1.15.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.13
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.22.0
tables==3.8.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.4
wrapt==1.14.1
xgboost==1.7.5
yarl==1.9.2
zipp==3.15.0
